Cast payload type to int in logs.
R=stefan@webrtc.org

Review URL: https://webrtc-codereview.appspot.com/29199004

git-svn-id: http://webrtc.googlecode.com/svn/trunk@7861 4adac7df-926f-26a2-2b94-8c16560cd09d

+++ webrtc/video_engine/vie_channel.cc
@@ -1800,8 +1800,8 @@ int32_t ViEChannel::OnInitializeDecoder(
     const int frequency,
     const uint8_t channels,
     const uint32_t rate) {
-  LOG(LS_INFO) << "OnInitializeDecoder " << payload_type << " "
-               << payload_name;
+  LOG(LS_INFO) << "OnInitializeDecoder " << static_cast<int>(payload_type)
+               << " " << payload_name;
   vcm_->ResetDecoder();
 
   CriticalSectionScoped cs(callback_cs_.get());
